Final Case Study Project - CON-E 330
This course is one of many that you will take towards your degree in Civil, Construction or Environmental Engineering. Each of our
courses are designed as part of your career development in your respective Engineering profession. Program Outcomes are intended to
provide a broad base of knowledge to find your career. However, each course in the curriculum emphasizes particular aspects of that
overall body of knowledge. Although other outcomes may also be addressed, this course is intended to have a particular emphasis on the
following program outcomes